Build #295 -- Version 0.0.0.18810 ---------------------------------
Directory trunk:

DEBUG.C:
  Temporarily avoid displaying debug messages.

DFNHDR.Y:
  Rename <DbgMsgW2> to <DbgMsgWP> and redefine the default behavior.

DISPDBG.C:
  Widen debug messages to use <DbgMsgW> in a move towards Unicode everywhere.

DISPLAY.C:
EDITFCN.C:
IMMEXEC.C:
MAIN.C:
YYFNS.C:
  Widen debug messages to use <DbgMsgW> in a move towards Unicode everywhere.

EXCEPT.C:
  Re-word leading text in Crash Control Window.
  Display the instruction stream.

EXCEDFN.C:
  Omit <DisplayFcnLine> in non-DEBUG versions.
  Use <DbgMsgW2> instead of <DbgMsgW> to avoid display of debug messages.

EXECMFN.C:
PF_COMMA.C:
PO_DOT.C:
QF_UCS.C:
SC_COPY.C:
SC_LOAD.C:
  Fix bug:  Check return value from functions that create a Symbol Table Entry.

MACROS.H:
  Define <DbgMsgW2> as controlled by the value of <gDbgLvl> to reduce debug messages.
  Use <nop> in <Assert> to allow easier continuation after error.
  Change <GetPtrTypeXXX> macro in DEBUG version to check for NULL ptr.
  Define <IsImmBool> macro to round out the set.

MAKEPRO.C:
  Handle empty output file.

PARSE.Y:
  Rename <DbgMsgW2> to <DbgMsgWP> and redefine the default behavior.
  Widen debug messages to use <DbgMsgW> in a move towards Unicode everywhere.

PF_EPSILON.C:
PO_DIERESIS.C:
  Fix bug:  Check return value from functions that create a Symbol Table Entry.
  Define <ERROR_EXIT> label and free result if allocated.

PF_INDEX.C:
PF_RSHOE.C:
PF_SQUAD.C:
  Fix bug:  Check return value from functions that create a Symbol Table Entry.
  Rename <ArrayIndexReplace> to <ArrayIndexReplace_EM> to reflect that
    it may return after signalling an error message.

PO_SLASH.C:
  Speed up monadic and dyadic reduction by determining the result storage type
    up front and blowing up only if necessary.  This avoids stressing the symbol
    table too much thus risking a SYMBOL TABLE FULL error message.
  Fix bug:  Dyadic reduction saves its values into the result in the wrong order
    if the reduction is not along the last coordinate.
  Fix bug:  Check return value from functions that create a Symbol Table Entry.

PO_SLOPE.C:
  Speed up monadic scan by determining the result storage type
    up front and blowing up only if necessary.  This avoids stressing the symbol
    table too much thus risking a SYMBOL TABLE FULL error message.
  Fix bug:  Check return value from functions that create a Symbol Table Entry.

PRIMSPEC.C:
  Improve comments.

SC_SAVE.C:
  Use macro for type testing.

SESSMAN.C:
  Widen debug messages to use <DbgMsgW> in a move towards Unicode everywhere.
  Fix bug:  Check return value from functions that create a Symbol Table Entry.

STARTADDR.H:
  Include more addresses for debugging purposes.

STRAND.C:
  In <MakeList_EM_YY>, move call to <YYAlloc> to after all errors are detected.
  In <MakeList_EM_YY>, fill the result with <PTR_REUSED> in case we fail part way through.

SYMTAB.C:
  Comment out <MakePermSymEntry_EM> as it not used.

TOKENIZE.C:
  Widen debug messages to use <DbgMsgW> in a move towards Unicode everywhere.

TYPEMOTE.C:
  Improve comments.
  Fix bug:  Check return value from functions that create a Symbol Table Entry.

+++ makepro.c
@@ -119,13 +119,13 @@ int PASCAL WinMain
                         lpTmpFile,
                         lpEndLine;
     SECURITY_ATTRIBUTES secAttr = {sizeof (SECURITY_ATTRIBUTES), NULL, TRUE};
-    HANDLE              hOutFile,
-                        hSrcFile,
-                        hTmpFile,
-                        hTmpMap,
-                        hOutMap,
-                        hTmpView,
-                        hOutView;
+    HANDLE              hOutFile = NULL,
+                        hSrcFile = NULL,
+                        hTmpFile = NULL,
+                        hTmpMap = NULL,
+                        hOutMap = NULL,
+                        hTmpView = NULL,
+                        hOutView = NULL;
     STARTUPINFO         siBlk;
     PROCESS_INFORMATION piBlk;
     DWORD               dwTmpFileSize,
@@ -261,6 +261,8 @@ int PASCAL WinMain
 
     // Get the file size (it always fits in a DWORD)
     dwTmpFileSize = GetFileSize (hTmpFile, NULL);
+    if (dwTmpFileSize EQ 0)
+        goto NORMAL_EXIT;
 
     // Now launder the output file to remove anomalies of some MS compilers
 
@@ -496,18 +498,27 @@ int PASCAL WinMain
         // We no longer need this resource
         CloseHandle (hOutFile); hOutFile = NULL;
     } // End IF
+NORMAL_EXIT:
+    if (hTmpView)
+    {
+        // We no longer need this resource
+        UnmapViewOfFile (hTmpView); hTmpView = NULL;
+    } // End IF
 
-    // We no longer need this resource
-    UnmapViewOfFile (hTmpView); hTmpView = NULL;
+    if (hTmpMap)
+    {
+        // We no longer need this resource
+        CloseHandle (hTmpMap); hTmpMap = NULL;
+    } // End IF
 
-    // We no longer need this resource
-    CloseHandle (hTmpMap); hTmpMap = NULL;
+    if (hTmpFile)
+    {
+        // We no longer need this resource
+        CloseHandle (hTmpFile); hTmpFile = NULL;
 
-    // We no longer need this resource
-    CloseHandle (hTmpFile); hTmpFile = NULL;
-
-    // We no longer need this resource
-    DeleteFile (lpTmpFile);
+        // We no longer need this resource
+        DeleteFile (lpTmpFile);
+    } // End IF
 
     return 0;
 } // End WinMain
